chore: Add new ServiceCommand type based on sealed classes
This removes the need for generics, and also gives more control over the definition of each service command. For example, this method means we don't need to add a `target` parameter to all service commands, just the ones that support it. Similarly, we can remove `serviceData` from commands that don't support it.