fix: Trigger Maven releases from a version tag
Previously we were waiting for the GitHub release workflow to finish in order to trigger the Maven release. Instead we'll trigger the Maven release based on the push of a version tag. The GitHub release workflow will push this tag, so we'll still trigger the Maven release, but now if the GitHub release doesn't happen then the Maven workflow isn't even started.