fix: Use command IDs to correlate responses
Previously we assumed that the next message after a websocket command request was a response to that request, but that's not a guarantee. This change uses the actual command IDs to correlate the response to the corresponding request. It uses Flows, which allows for multiple collectors and therefore can process multiple commands simultaneously.